requirements.txt
asgiref==3.8.1
colorama==0.4.6
crayons==0.4.0
Django==5.1.7
django-cors-headers==4.7.0
numpy==2.2.3
opencv-python==4.11.0.86
piexif==1.1.3
pillow==10.4.0
sqlparse==0.5.3
stegano==0.11.5
tzdata==2025.1
